feature:单位管理功能开发

+++ src/app/manage/system/unit-manage/unit-manage.component.ts
@@ -22,9 +22,8 @@ export class UnitManageComponent implements OnInit {
   listOfAllData:params[] = []
   mapOfCheckedId: { [key: string]: boolean } = {};
   selectItems = []
-  baseUrl='/account'
-  checkUrl:'/account/enable' //启用
-  stopUrl:'/account/stop' //停用
+  baseUrl='/unit'
+  submitAudit='/unit/submitAudit'
 
   constructor(
     private modal: NzModalService,
@@ -56,18 +55,7 @@ export class UnitManageComponent implements OnInit {
   }
   //新增操作
   Add() {
-    const param = {
-      account:"",
-      pass:"",
-      name:"",
-      unit_id:'',
-      phone:'',
-      email:'',
-      on_trial:'',
-      trial_end:'',
-      role_id:'',
-      is_super:'',
-    }
+    const param = new params;
     let modalRef:NzModalRef = this.modal.create({
       nzTitle:"参数配置",
       nzContent:ManageFormComponent,
@@ -84,7 +72,6 @@ export class UnitManageComponent implements OnInit {
           disabled:comp => !comp.validateForm.valid,
           onClick:(comp) => {
             let formVal = comp.validateForm.getRawValue()
-            console.log(formVal);
             this.modal.confirm({
               nzTitle: '提交',
               nzContent: '确认提交?',
@@ -114,6 +101,7 @@ export class UnitManageComponent implements OnInit {
       this.msg.warning('请选择一项数据进行操作!')
       return;
     }
+    if(this.selectItems[0].state === '待审核' || this.selectItems[0].state ==='审核通过') {   this.msg.warning("请选择状态为'草稿'或'审核不通过'的数据项进行操作!");return;}
     let modalRef:NzModalRef = this.modal.create({
       nzTitle:"参数配置",
       nzContent:ManageFormComponent,
@@ -162,6 +150,8 @@ export class UnitManageComponent implements OnInit {
       this.msg.warning('请先选择数据进行操作!')
       return;
     }
+    const flag = this.selectItems.every(item => item.state !== '待审核' && item.state !== '审核通过')
+    if( !flag) {this.msg.warning("请选择状态为'草稿'或'审核不通过'的数据项进行操作!"); return;}
     const selectedIds = this.selectItems.map(it => it.id) + ''
     this.http.delete(`${this.baseUrl}/${selectedIds}`).subscribe(res => {
       if(res.code !==0) {
@@ -172,19 +162,19 @@ export class UnitManageComponent implements OnInit {
       this.getData();
     })
   }
-  //启用
-  Check() {
+  //提交审核
+  submitCheck() {
     if(this.selectItems.length === 0 ) {
       this.msg.warning('请先选择数据进行操作!')
       return;
     }
-    const checkStatus = this.selectItems.every(it => it.enable)
-    if(!checkStatus) {
-      this.msg.warning('请选择禁用状态的数据进行操作')
+    const flag = this.selectItems.every(it => it.state === '草稿')
+    if(!flag) {
+      this.msg.warning('请选择草稿状态的数据项进行操作')
       return;
     }
     const selectedIds = this.selectItems.map(it => it.id) + ''
-    this.http.post(`${this.checkUrl}?ids=${selectedIds}`).subscribe(res => {
+    this.http.get(`${this.submitAudit}?ids=${selectedIds}`).subscribe(res => {
       if(res.code !== 0) {
         this.msg.error(res.message);
         return
@@ -193,27 +183,7 @@ export class UnitManageComponent implements OnInit {
       this.getData();
     })
   }
-  //停用
-  Stop () {
-    if(this.selectItems.length === 0 ) {
-      this.msg.warning('请先选择数据进行操作!')
-      return;
-    }
-    const checkStatus = this.selectItems.every(it => !it.enable)
-    if(!checkStatus) {
-      this.msg.warning('请选择启用状态的数据进行操作')
-      return;
-    }
-    const selectedIds = this.selectItems.map(it => it.id) + ''
-    this.http.post(`${this.stopUrl}?ids=${selectedIds}`).subscribe(res => {
-      if(res.code !== 0) {
-        this.msg.error(res.message);
-        return
-      }
-      this.msg.success(res.message);
-      this.getData();
-    })
-  }
+
   //刷新
   Query() {
     this.getData()
@@ -247,16 +217,21 @@ export class UnitManageComponent implements OnInit {
 }
 class params {
   id:number        //编号
-  account:string;  //账号
-  pass:string;     //登录密码
-  name:string;    //姓名
-  unit_id:number;  //所属单位
-  phone:string;     //联系电话
-  email:string;     //邮箱
-  enable:boolean;  //是否启用
-  on_trial:boolean;  //是否试用
-  trial_end	:string;  //试用到期时间
-  role_id:string;  //角色
-  super:boolean;  //是否为管理员
-  manage_host?:string;  //管理主机授权
+  unit_name:string;  //单位名称
+  social_code:string;     //统一社会信用代码
+  parent_id:string;    //上级单位
+  unit_type:number;  //单位类型
+  state:string;     //状态
+  fixed_phone:string;     //单位固话
+  linkman:boolean;  //联系人
+  cell_phone:boolean;  //联系人手机
+  unit_address	:string;  //单位地址
+  unit_email:string;  //邮箱
+  fax:boolean;  //传真
+  bank?:string;  //开户银行
+  bank_account?:string;  //银行账号
+  auditor?:string;  //审核人
+  audit_mark?:string;  //审核备注
+  audit_time?:string;  //审核时间
+  mark?:string;  //备注
 }
